refactor example
