--- examples/groth16-verifier/groth16/Cargo.toml
@@ -1,18 +1,0 @@
-[package]
-name = "risc0-groth16"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-anyhow = "1"
-ark-bn254 = "0.4"
-ark-groth16 = "0.4"
-ark-serialize = "0.4"
-bincode = "1.3"
-bytemuck = "1.14"
-ethereum-types = "0.14"
-hex = "0.4"
-risc0-zkvm = { path = "../../../risc0/zkvm", default-features = false }
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
-sha2 = { version = "=0.10.6" }
