+++ examples/groth16-verifier/methods/guest/Cargo.toml
@@ -6,7 +6,7 @@ edition = "2021"
 [workspace]
 
 [dependencies]
-risc0-groth16 = { path = "../../groth16" }
+risc0-groth16 = { path = "../../../../risc0/groth16", default-features = false }
 risc0-zkvm = { path = "../../../../risc0/zkvm", default-features = false, features = ["std"] }
 
 [patch.crates-io]
