+++ examples/groth16-verifier/methods/guest/src/bin/groth16_verifier.rs
@@ -14,13 +14,13 @@
 
 #![no_main]
 
-use risc0_groth16::Groth16;
+use risc0_groth16::Verifier;
 use risc0_zkvm::guest::env;
 
 risc0_zkvm::guest::entry!(main);
 
 pub fn main() {
-    let proof: Groth16 = env::read();
-    proof.verify().unwrap();
-    env::commit(&proof.digest());
+    let verifier: Verifier = env::read();
+    verifier.verify().unwrap();
+    // env::commit(&proof.digest());
 }
