+++ examples/groth16-verifier/src/data/proof.json
