+++ examples/groth16-verifier/src/data/public.json
