+++ examples/groth16-verifier/src/data/verification_key.json
