+++ examples/groth16-verifier/src/main.rs
@@ -13,34 +13,31 @@
 // limitations under the License.
 
 use methods::{GROTH16_VERIFIER_ELF, GROTH16_VERIFIER_ID};
-use risc0_groth16::{
-    circom::{CircomProof, CircomPublic, CircomVKey},
-    Groth16,
-};
+use risc0_groth16::{ProofJson, PublicInputsJson, Verifier, VerifyingKeyJson};
 use risc0_zkvm::{default_prover, ExecutorEnv};
 
-const CIRCOM_VERIFICATION_KEY: &str = include_str!("data/circom/verification_key.json");
-const CIRCOM_PROOF: &str = include_str!("data/circom/proof.json");
-const CIRCOM_PUBLIC: &str = include_str!("data/circom/public.json");
+const PROOF: &str = include_str!("data/proof.json");
+const PUBLIC_INPUTS: &str = include_str!("data/public.json");
+const VERIFICATION_KEY: &str = include_str!("data/verification_key.json");
 
 fn main() {
-    // verification_key, proof and public witness generated with circom and snarkjs using Groth16 over BN254
+    // verification_key, proof and public witness generated by SnarkJS using Groth16 over BN254
     // (https://docs.circom.io/getting-started/proving-circuits/)
-    let circom_vkey: CircomVKey = serde_json::from_str(CIRCOM_VERIFICATION_KEY).unwrap();
-    let circom_proof: CircomProof = serde_json::from_str(CIRCOM_PROOF).unwrap();
-    let circom_public = CircomPublic {
-        values: serde_json::from_str(CIRCOM_PUBLIC).unwrap(),
+    let proof: ProofJson = serde_json::from_str(PROOF).unwrap();
+    let public_inputs = PublicInputsJson {
+        values: serde_json::from_str(PUBLIC_INPUTS).unwrap(),
     };
+    let verifying_key: VerifyingKeyJson = serde_json::from_str(VERIFICATION_KEY).unwrap();
 
-    // we parse a groth16 proof from the circom material collected from SnarkJS
-    let groth16_proof = Groth16::from_circom(circom_vkey, circom_proof, circom_public).unwrap();
+    // we initialize a Groth16 verifier from the josn material generated by SnarkJS
+    let verifier = Verifier::from_json(proof, public_inputs, verifying_key).unwrap();
 
     // groth16 proof verification
-    groth16_proof.verify().unwrap();
+    verifier.verify().unwrap();
 
-    // we configure an Executor with the groth16 proof
+    // we configure an Executor with the groth16 verifier
     let env = ExecutorEnv::builder()
-        .write(&groth16_proof)
+        .write(&verifier)
         .unwrap()
         .build()
         .unwrap();
@@ -50,10 +47,6 @@ fn main() {
 
     // we verify the final receipt
     receipt.verify(GROTH16_VERIFIER_ID).unwrap();
-
-    // we compare the committed hash to the original proof
-    let output_digest: [u8; 32] = receipt.journal.decode().unwrap();
-    assert_eq!(output_digest, groth16_proof.digest());
 
     println!("Verification: OK!");
 }
