+++ risc0/groth16/Cargo.toml
@@ -22,6 +22,7 @@ serde = { version = "1.0", default-features = false, features = [
 ] }
 
 [dev-dependencies]
+serde_json = "1.0"
 
 [features]
 std = [
