+++ risc0/groth16/tests/data/verification_key.json
